{"schema":"decodex/commit/1","summary":"Split live consolidation adapter modules","authority":"manual"}
